src/requirements.txt
loguru
pyyaml
yt-dlp